requirements.txt
google-cloud-documentai>=2.30.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1
